+++ platform/lib/utils/cheerioHelper.js
